Add basic contact page.
Needs to be fleshed out more.

The rationale for having a dedicated page is that people trying to
find contact info quickly should be able to so easily, and some
people might not notice the contact info in the footer.

Also, having a dedicated contact page also allows us to briefly
describe how and for what the club uses each form of communication.