Drop site.baseurl everywhere.
It was already unmaintained, and was just going to get worse.

If we really need such a feature later on we'd need to audit the
whole site at that point anyway. Lets just save ourselves the
trouble of trying to maintain untested features until then.

We'll likely never want to use it anyway.